+++ src/game_object/player/ClassicFighter.java
@@ -28,12 +28,11 @@ public class ClassicFighter extends Character
 
     /**
      * Constructing the character with given parameters.
-     *
-     * @param x                 - x coordinate of the character.
+     *  @param x                 - x coordinate of the character.
      * @param y                 - y coordinate of the character.
      * @param id                - id of the character as a game object.
      */
-    public ClassicFighter(float x, float y, ObjectID id)
+    public ClassicFighter(double x, double y, ObjectID id)
     {
         super(x, y, id);
